+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * [FIX] Make icon scale deserialization work also with integers, not only doubles
 * [ADD] Add a new drawing for colored rectangles which allows for partially clearing the screen
 * [FIX] Fix blank.svg not rendering
+* [ADD] Add `SpinnerDrawing` for drawing animated spinners
 
 ## 1.1.3
 * [DEL] Deprecate delays when sending commands and set their default value to 0